generic temperature sensor component
